// teams/people/organizations-management/base/src/domains/users-v1/core/entity.schema.types.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface IUserV1Dto {
  /**
   * The unique identifier for a user, expected to be 28 characters long.
   */
  id: string;
  /**
   * The email address of the user.
   */
  email: string;
  /**
   * Flag indicating whether the user's email address has been verified.
   */
  emailVerified: boolean;
  /**
   * The display name of the user.
   */
  displayName: string;
  /**
   * The URL of the user's photo.
   */
  photoURL?: string | null;
  /**
   * Flag indicating whether the user account is disabled.
   */
  disabled: boolean;
  /**
   * The creation date and time of the user account.
   */
  createdAt: string;
  /**
   * The last update date and time of the user account.
   */
  updatedAt: string;
}